requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --no-index --output-file requirements.txt requirements.in
#
aiohttp==3.4.4
asn1crypto==0.24.0        # via cryptography
async-timeout==3.0.1      # via aiohttp
attrs==18.2.0             # via aiohttp
botbuilder-core==4.0.0a6
botbuilder-schema==4.0.0a6  # via botbuilder-core, botframework-connector
botframework-connector==4.0.0a6  # via botbuilder-core
certifi==2018.10.15       # via msrest, requests
cffi==1.11.5              # via cryptography
chardet==3.0.4            # via aiohttp, requests
cryptography==2.3.1       # via botframework-connector
idna==2.7                 # via cryptography, requests, yarl
iso8601==0.1.12
isodate==0.6.0            # via msrest
msrest==0.4.29            # via botbuilder-schema, botframework-connector
multidict==4.4.2          # via aiohttp, yarl
oauthlib==2.1.0           # via requests-oauthlib
pycparser==2.19           # via cffi
pycrypto==2.6.1
pyjwt==1.6.4              # via botframework-connector
requests-oauthlib==1.0.0  # via msrest
requests==2.20.0          # via botframework-connector, msrest, requests-oauthlib
six==1.11.0               # via cryptography, isodate
urllib3==1.24             # via requests
yarl==1.2.6               # via aiohttp
